+++ ITMO Stars/Assets/Scripts/Scriptable/EnemyInfo.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Mobs/New Enemy")]
 public class EnemyInfo : MobInfo
 {
-    public float AttackStrength;
+    public int AttackStrength;
 }
